refactor(commentaries): extract pager/grid scaffolding to CommentatorsGrid.kt
Moves grid constants, capacity/layout helpers, and pager composables
(CommentatorsGridScaffold, VerticalPagerIndicator, CommentatorHeader, etc.)
to a dedicated file for better maintainability. Changes CommentariesLayoutConfig
and AnimatedTextSizes to internal visibility for cross-file access. Merges
two consecutive LaunchedEffect(pagerState.settledPage) into a single effect
that notifies the caller and updates the anchor in one go.
